✨ Add Projects Write Route and Refactor Dual Sidebar for Enhanced Writing Experience ✨
- Introduced a new route for writing projects at `/projects/$projectId/write`, enhancing the project management capabilities.
- Refactored the `DualSidebarWritingInterface` to integrate the new `ProjectSidebar`, streamlining the sidebar structure and improving user navigation.
- Removed unused imports and cleaned up the codebase for better maintainability.

Kawaii updates for a more organized and efficient writing journey, senpai! 💖
