+++ apps/web/src/components/dual-sidebar-writing-interface.tsx
@@ -1,42 +1,23 @@
 import { useQuery } from "@tanstack/react-query"
 import { Link } from "@tanstack/react-router"
-import { ChevronDown, ChevronRight, FileText, PenTool, Plus, Scroll, Sparkles } from "lucide-react"
+import { Sparkles } from "lucide-react"
 import { useState } from "react"
 import { useHotkeys } from "react-hotkeys-hook"
 import { AIChatContent } from "@/components/ai-chat-content"
 import { AutocompleteToggle } from "@/components/autocomplete-toggle"
-import { CharacterSidebarSection } from "@/components/character-sidebar-section"
-import CodexModal from "@/components/codex-modal"
-import { LocationSidebarSection } from "@/components/location-sidebar-section"
+import { ProjectSidebar } from "@/components/project-sidebar"
 import TiptapEditor from "@/components/tiptap-editor"
-import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
-import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@/components/ui/collapsible"
-import { Logo } from "@/components/ui/logo"
 import {
   Sheet,
   SheetContent,
   SheetDescription,
   SheetHeader,
   SheetTitle,
+  SheetTrigger,
 } from "@/components/ui/sheet"
-import {
-  Sidebar,
-  SidebarContent,
-  SidebarFooter,
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarGroupLabel,
-  SidebarHeader,
-  SidebarInset,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem,
-  SidebarProvider,
-  SidebarTrigger,
-} from "@/components/ui/sidebar"
+import { SidebarInset, SidebarProvider, SidebarTrigger } from "@/components/ui/sidebar"
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip"
-import UserMenu from "@/components/user-menu"
 import { useIsMobile } from "@/hooks/use-mobile"
 import { api } from "@/lib/api"
 
@@ -54,19 +35,6 @@ function DualSidebarWritingInterfaceInner({
   onUpdate,
   placeholder = "Write something beautiful...",
 }: DualSidebarWritingInterfaceProps) {
-  const [isCodexModalOpen, setIsCodexModalOpen] = useState(false)
-  const [codexModalConfig, setCodexModalConfig] = useState<{
-    initialType?: string | null
-    initialEntry?: string | null
-  }>({})
-  const [expandedCodexSections, setExpandedCodexSections] = useState<Record<string, boolean>>({
-    characters: false,
-    locations: false,
-    lore: false,
-    plot: false,
-    notes: false,
-  })
-
   // Separate state for right sidebar (AI chat) - independent from left sidebar
   const [rightSidebarOpen, setRightSidebarOpen] = useState(true)
   const isMobile = useIsMobile()
@@ -116,19 +84,6 @@ function DualSidebarWritingInterfaceInner({
     ? Math.min((project.currentWordCount / targetWordCount) * 100, 100)
     : 0
 
-  const toggleCodexSection = (section: string) => {
-    console.log(
-      `🎯 toggleCodexSection("${section}"):`,
-      expandedCodexSections[section],
-      "->",
-      !expandedCodexSections[section]
-    )
-    setExpandedCodexSections((prev) => ({
-      ...prev,
-      [section]: !prev[section],
-    }))
-  }
-
   const handleInsertText = (text: string) => {
     // This would integrate with the TiptapEditor to insert text at cursor position
     if (onUpdate) {
@@ -136,259 +91,10 @@ function DualSidebarWritingInterfaceInner({
     }
   }
 
-  // Locations data will be fetched from API via LocationSidebarSection
-
-  const loreEntries: Array<{
-    id: string
-    name: string
-    type: string
-    description?: string
-  }> = []
-
-  const plotThreads: Array<{
-    name: string
-    role: string
-    description?: string
-  }> = []
-
-  const notes: Array<{
-    name: string
-    role: string
-    description?: string
-  }> = []
-
-  const openCodexModal = (type?: string, entry?: string) => {
-    setCodexModalConfig({
-      initialType: type || null,
-      initialEntry: entry || null,
-    })
-    setIsCodexModalOpen(true)
-  }
-
-  const closeCodexModal = () => {
-    setIsCodexModalOpen(false)
-    setCodexModalConfig({})
-  }
-
   return (
     <>
-      {/* Left Sidebar - Navigation & Codex */}
-      <Sidebar side="left" variant="sidebar">
-        <SidebarHeader>
-          <div className="px-4 py-3">
-            <Logo size="md" to="/dashboard" />
-          </div>
-        </SidebarHeader>
-
-        <SidebarContent>
-          <SidebarGroup>
-            <SidebarGroupLabel>Structure</SidebarGroupLabel>
-            <SidebarGroupContent>
-              <SidebarMenu>
-                <SidebarMenuItem>
-                  <SidebarMenuButton asChild>
-                    <Link params={{ projectId }} to="/projects/$projectId/canvas">
-                      <FileText />
-                      <span>Canvas</span>
-                    </Link>
-                  </SidebarMenuButton>
-                </SidebarMenuItem>
-                <SidebarMenuItem>
-                  <SidebarMenuButton asChild>
-                    <Link params={{ projectId }} to="/write/$projectId">
-                      <PenTool />
-                      <span>Write</span>
-                    </Link>
-                  </SidebarMenuButton>
-                </SidebarMenuItem>
-              </SidebarMenu>
-            </SidebarGroupContent>
-          </SidebarGroup>
-
-          <SidebarGroup>
-            <SidebarGroupLabel>Codex</SidebarGroupLabel>
-            <SidebarGroupContent>
-              <SidebarMenu>
-                {/* Characters */}
-                <CharacterSidebarSection
-                  isExpanded={expandedCodexSections.characters}
-                  key="characters-section"
-                  onOpenCodexModal={openCodexModal}
-                  onToggle={() => toggleCodexSection("characters")}
-                  projectId={projectId}
-                />
-
-                {/* Locations */}
-                <LocationSidebarSection
-                  isExpanded={expandedCodexSections.locations}
-                  key="locations-section"
-                  onOpenCodexModal={openCodexModal}
-                  onToggle={() => toggleCodexSection("locations")}
-                  projectId={projectId}
-                />
-
-                {/* Lore */}
-                <SidebarMenuItem>
-                  <Collapsible
-                    onOpenChange={() => toggleCodexSection("lore")}
-                    open={expandedCodexSections.lore}
-                  >
-                    <CollapsibleTrigger asChild>
-                      <SidebarMenuButton>
-                        {expandedCodexSections.lore ? (
-                          <ChevronDown className="h-4 w-4" />
-                        ) : (
-                          <ChevronRight className="h-4 w-4" />
-                        )}
-                        <Scroll className="h-4 w-4" />
-                        <span>Lore</span>
-                        <Badge className="ml-auto" variant="secondary">
-                          {loreEntries.length}
-                        </Badge>
-                      </SidebarMenuButton>
-                    </CollapsibleTrigger>
-                    <CollapsibleContent>
-                      <div className="ml-6 space-y-1">
-                        <Button
-                          className="w-full justify-start text-muted-foreground"
-                          onClick={() => openCodexModal("lore")}
-                          size="sm"
-                          variant="ghost"
-                        >
-                          <Plus className="h-4 w-4" />
-                          <span>New</span>
-                        </Button>
-                        {loreEntries.map((loreItem) => (
-                          <Button
-                            className="w-full justify-start"
-                            key={loreItem.id}
-                            onClick={() => openCodexModal("lore", loreItem.name)}
-                            size="sm"
-                            variant="ghost"
-                          >
-                            <span className="truncate">{loreItem.name}</span>
-                            <span className="ml-auto text-muted-foreground text-xs">
-                              {loreItem.type}
-                            </span>
-                          </Button>
-                        ))}
-                      </div>
-                    </CollapsibleContent>
-                  </Collapsible>
-                </SidebarMenuItem>
-
-                {/* Plot Threads */}
-                <SidebarMenuItem>
-                  <Collapsible
-                    onOpenChange={() => toggleCodexSection("plot")}
-                    open={expandedCodexSections.plot}
-                  >
-                    <CollapsibleTrigger asChild>
-                      <SidebarMenuButton>
-                        {expandedCodexSections.plot ? (
-                          <ChevronDown className="h-4 w-4" />
-                        ) : (
-                          <ChevronRight className="h-4 w-4" />
-                        )}
-                        <FileText className="h-4 w-4" />
-                        <span>Plot Threads</span>
-                        <Badge className="ml-auto" variant="secondary">
-                          {plotThreads.length}
-                        </Badge>
-                      </SidebarMenuButton>
-                    </CollapsibleTrigger>
-                    <CollapsibleContent>
-                      <div className="ml-6 space-y-1">
-                        <Button
-                          className="w-full justify-start text-muted-foreground"
-                          onClick={() => openCodexModal("plot")}
-                          size="sm"
-                          variant="ghost"
-                        >
-                          <Plus className="h-4 w-4" />
-                          <span>New</span>
-                        </Button>
-                        {plotThreads.map((plotItem) => (
-                          <Button
-                            className="w-full justify-start"
-                            key={plotItem.name}
-                            onClick={() => openCodexModal("plot", plotItem.name)}
-                            size="sm"
-                            variant="ghost"
-                          >
-                            <span className="truncate">{plotItem.name}</span>
-                            <span className="ml-auto text-muted-foreground text-xs">
-                              {plotItem.role}
-                            </span>
-                          </Button>
-                        ))}
-                      </div>
-                    </CollapsibleContent>
-                  </Collapsible>
-                </SidebarMenuItem>
-
-                {/* Notes */}
-                <SidebarMenuItem>
-                  <Collapsible
-                    onOpenChange={() => toggleCodexSection("notes")}
-                    open={expandedCodexSections.notes}
-                  >
-                    <CollapsibleTrigger asChild>
-                      <SidebarMenuButton>
-                        {expandedCodexSections.notes ? (
-                          <ChevronDown className="h-4 w-4" />
-                        ) : (
-                          <ChevronRight className="h-4 w-4" />
-                        )}
-                        <FileText className="h-4 w-4" />
-                        <span>Notes & Ideas</span>
-                        <Badge className="ml-auto" variant="secondary">
-                          {notes.length}
-                        </Badge>
-                      </SidebarMenuButton>
-                    </CollapsibleTrigger>
-                    <CollapsibleContent>
-                      <div className="ml-6 space-y-1">
-                        <Button
-                          className="w-full justify-start text-muted-foreground"
-                          onClick={() => openCodexModal("notes")}
-                          size="sm"
-                          variant="ghost"
-                        >
-                          <Plus className="h-4 w-4" />
-                          <span>New</span>
-                        </Button>
-                        {notes.map((note) => (
-                          <Button
-                            className="w-full justify-start"
-                            key={note.name}
-                            onClick={() => openCodexModal("notes", note.name)}
-                            size="sm"
-                            variant="ghost"
-                          >
-                            <span className="truncate">{note.name}</span>
-                            <span className="ml-auto text-muted-foreground text-xs">
-                              {note.role}
-                            </span>
-                          </Button>
-                        ))}
-                      </div>
-                    </CollapsibleContent>
-                  </Collapsible>
-                </SidebarMenuItem>
-              </SidebarMenu>
-            </SidebarGroupContent>
-          </SidebarGroup>
-        </SidebarContent>
-
-        <SidebarFooter>
-          <SidebarMenu>
-            <SidebarMenuItem>
-              <UserMenu />
-            </SidebarMenuItem>
-          </SidebarMenu>
-        </SidebarFooter>
-      </Sidebar>
+      {/* Project Sidebar */}
+      <ProjectSidebar projectId={projectId} />
 
       {/* Main Content Area */}
       <SidebarInset className="flex flex-1 flex-col">
@@ -501,15 +207,6 @@ function DualSidebarWritingInterfaceInner({
           </div>
         )
       )}
-
-      {/* Codex Modal */}
-      <CodexModal
-        initialEntry={codexModalConfig.initialEntry}
-        initialType={codexModalConfig.initialType}
-        isOpen={isCodexModalOpen}
-        onClose={closeCodexModal}
-        projectId={projectId}
-      />
     </>
   )
 }
