+++ apps/web/src/routes/projects/$projectId/canvas.tsx
@@ -62,6 +62,7 @@ import {
   SheetHeader,
   SheetTitle,
 } from "@/components/ui/sheet"
+
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Textarea } from "@/components/ui/textarea"
